refactor(ao): extract orchestration select adapter
