+++ docs/reduction/ao-file-buckets.tsv
@@ -368,8 +368,8 @@ cli/cmd/ao/operator_adapter.go	KEEP	self-contained AO image surface: loop, mind,
 cli/cmd/ao/operator_adapter_test.go	KEEP	self-contained AO image surface: loop, mind, local Themis, local flywheel, corpus, or operator ergonomics	context,errors,internal/ports,os,path/...,strings,testing
 cli/cmd/ao/operator_cmd.go	KEEP	self-contained AO image surface: loop, mind, local Themis, local flywheel, corpus, or operator ergonomics	context,encoding/...,errors,fmt,github.com/...,internal/ports,io,os,path/...
 cli/cmd/ao/operator_cmd_test.go	KEEP	self-contained AO image surface: loop, mind, local Themis, local flywheel, corpus, or operator ergonomics	bytes,context,errors,internal/ports,strings,testing
-cli/cmd/ao/orchestrate.go	RELOCATE	outer orchestration, runtime supervision, or fleet/adapter control; move toward MTO or vendor-image adapter boundary	context,encoding/...,fmt,github.com/...,internal/orchestration,internal/ports,os/...,strings
-cli/cmd/ao/orchestrate_test.go	RELOCATE	outer orchestration, runtime supervision, or fleet/adapter control; move toward MTO or vendor-image adapter boundary	bytes,context,encoding/...,errors,github.com/...,internal/orchestration,internal/ports,strings,testing
+cli/cmd/ao/orchestrate.go	KEEP	public compatibility wrapper for `ao orchestrate select`; production selector moved to the mto-fleet route	fmt,github.com/...,internal/adapters
+cli/cmd/ao/orchestrate_test.go	KEEP	wrapper registration test for the retained `ao orchestrate select` public command surface	testing
 cli/cmd/ao/parity_test.go	KEEP	test coverage follows the command surface under review; keep while the source file remains	encoding/...,fmt,internal/goals,internal/pool,internal/types,internal/vibecheck,os,path/...,strings,testing,time
 cli/cmd/ao/patterns_repair.go	KEEP	self-contained AO image surface: loop, mind, local Themis, local flywheel, corpus, or operator ergonomics	bufio,fmt,github.com/...,io,os,path/...,regexp,sort,strings
 cli/cmd/ao/patterns_repair_test.go	KEEP	self-contained AO image surface: loop, mind, local Themis, local flywheel, corpus, or operator ergonomics	bytes,os,path/...,sort,strings,testing
